Changed method isUnique to accept unicode and tests added.
The method isUnique is now using a integer (32 bit) as a checker
and it is put in end with a 1 shifted val times, where val is the
value of the char. If && is positive means that the same value was
added to the checker previously. To update the checker, the new 1 shifted
val times is added with an or (|) operation.
This new version of is Unique is able to check any character.